Move input and payload types from user to schema layer
