+++ src/graphql/schema/types/user.ts
@@ -1,4 +1,4 @@
-import { createUnionType, Field, ID, ObjectType } from 'type-graphql'
+import { createUnionType, Field, ID, InputType, ObjectType } from 'type-graphql'
 import { ErrorCode } from 'graphql/schema/enums/errorCode'
 import { Post } from './post'
 import { UserError } from './userError'
@@ -48,4 +48,27 @@ export class UserNameTakenError implements Partial<UserError> {
 export const CreateUserPayload = createUnionType({
   name: 'CreateUserPayload',
   types: () => [CreateUserSuccess, EmailTakenError, UserNameTakenError] as const
+})
+
+@InputType()
+export class CreateUserInput implements Partial<User> {
+  @Field()
+    name: string
+
+  @Field()
+    email: string
+}
+@InputType()
+export class DeleteUserInput implements Partial<User> {
+  @Field(() => ID)
+    id: string
+}
+@ObjectType()
+export class DeleteUserSuccess {
+  @Field(() => User)
+    user: User
+}
+export const DeleteUserPayload = createUnionType({
+  name: 'DeleteUserPayload',
+  types: () => [DeleteUserSuccess, UserError] as const
 })
